feat: update deleteItem mutation to check required permissions

+++ backend/src/resolvers/Mutation.js
@@ -50,9 +50,15 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1 Find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id} }`);
     // 2 Check if the user owns the item or has a permission
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error(`You don't have permission to do that!`);
+    }
     // 3 Delete the item
     return ctx.db.mutation.deleteItem({ where }, info);
   },
